Wrote script for importing courses to firestore
Script.ts contains a import courses function and a csv reader to take courses from a csv file (in the website_data_csv folder), parse the information into a list of JSON objects, and create a doc for each object in the courses collection in firestore. Also created a command "yarn script" in the package.json to run this script.